Tier-2: HSLM float32 under AMP, atomic result writes, code-rev stamp
- 4.1  HSLM marginalization forces logits.float() before softmax/log. It runs
       inside AMP; in fp16 a small species prob underflows to 0 -> log(0)=-inf
       and the genus/family loss explodes, and eps=1e-8 is below fp16's smallest
       normal so the clamp can't save it. float32 makes softmax+log safe.
       Verified end-to-end under autocast: backward works, components sane.
- 2.6  ResultWriter writes provenance + checkpoint FIRST and metrics.json LAST,
       each via a temp-file + os.replace() atomic rename. already_done() keys on
       metrics.json, so its existence now means the whole dir is complete; a
       crash mid-save leaves no metrics.json and the run re-executes cleanly
       instead of being skipped half-written. train.py shares the writer.
- 2.7  each metrics.json records the short git commit it was produced at, so a
       result carried across a code change is at least detectable.

+++ bioreef/model/hslm_loss.py
@@ -116,8 +116,12 @@ class HSLMLoss(nn.Module):
         w = self.cb_weights[targets]
         species_loss = (w * (1.0 - pt) ** self.gamma * ce).mean()
 
-        # Marginalize species probabilities up the taxonomy.
-        p_species = F.softmax(logits, dim=1)
+        # Marginalize species probabilities up the taxonomy. Force float32: this
+        # runs inside AMP, and in fp16 a small species prob can underflow to 0,
+        # then log() -> -inf and the genus/family loss explodes. eps=1e-8 is also
+        # below fp16's smallest normal (~6e-5), so the clamp wouldn't even save
+        # it. In float32 both the softmax and the log are safe.
+        p_species = F.softmax(logits.float(), dim=1)
         p_genus = self._marginalize(p_species, self.species_to_genus, self.num_genera)
         p_family = self._marginalize(p_species, self.species_to_family, self.num_families)
